adding feature and committing the changes

+++ Backend Services/src/main/java/com/restaurant/backend/payloads/ItemOrderDTO.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Getter
 public class ItemOrderDTO extends BaseDTO {
     private long id;
-    private double quantity;
+    private int quantity;
     private double price;
 
     // it takes only id of item and order instead of complete objects when order would be placed
